+++ pdfminer/high_level.py
@@ -1,14 +1,30 @@
 """Functions that can be used for the most common use-cases for pdfminer.six"""
-
+ 
+import logging
+import six
 import sys
 
-import six
-
+# Conditional import because python 2 is stupid
+if sys.version_info > (3, 0):
+    from io import StringIO
+else:
+    from io import BytesIO as StringIO
+    
 from pdfminer.converter import XMLConverter, HTMLConverter, TextConverter
 from pdfminer.image import ImageWriter
 from pdfminer.pdfdevice import TagExtractor
 from pdfminer.pdfinterp import PDFResourceManager, PDFPageInterpreter
 from pdfminer.pdfpage import PDFPage
+
+from .pdfdocument import PDFDocument
+from .pdfparser import PDFParser
+from .pdfinterp import PDFResourceManager, PDFPageInterpreter
+from .pdfdevice import PDFDevice, TagExtractor
+from .pdfpage import PDFPage
+from .converter import XMLConverter, HTMLConverter, TextConverter
+from .cmapdb import CMapDB
+from .image import ImageWriter
+from .layout import LAParams
 
 
 def extract_text_to_fp(inf, outfp,
@@ -48,6 +64,9 @@ def extract_text_to_fp(inf, outfp,
             'upgrade to Python 3. For more information see '
             'https://github.com/pdfminer/pdfminer .six/issues/194')
 
+    if debug:
+        logging.getLogger().setLevel(logging.DEBUG)
+
     if six.PY2 and sys.stdin.encoding:
         password = password.decode(sys.stdin.encoding)
 
@@ -86,3 +105,41 @@ def extract_text_to_fp(inf, outfp,
         interpreter.process_page(page)
 
     device.close()
+
+
+def extract_text(pdf_file, password='', page_numbers=None, maxpages=0,
+                 caching=True, codec='utf-8', laparams=None):
+    """
+    Parses and returns the text contained in a PDF file.
+    Takes loads of optional arguments but the defaults are somewhat sane.
+    Returns a string containing all of the text extracted.
+
+    :param pdf_file: Path to the PDF file to be worked on
+    :param password: For encrypted PDFs, the password to decrypt.
+    :param page_numbers: List of zero-indexed page numbers to extract.
+    :param maxpages: The maximum number of pages to parse
+    :param caching: If resources should be cached
+    :param codec: Text decoding codec
+    :param laparams: LAParams object from pdfminer.layout.
+    """
+    if laparams is None:
+        laparams = LAParams()
+
+    with open(pdf_file, "rb") as fp, StringIO() as output_string:
+        rsrcmgr = PDFResourceManager()
+        device = TextConverter(rsrcmgr, output_string, codec=codec,
+                               laparams=laparams)
+        interpreter = PDFPageInterpreter(rsrcmgr, device)
+
+        for page in PDFPage.get_pages(
+            fp,
+            page_numbers,
+            maxpages=maxpages,
+            password=password,
+            caching=caching,
+            check_extractable=True,
+        ):
+            interpreter.process_page(page)
+
+        return output_string.getvalue()
+