+++ tools/pdf2txt.py
@@ -1,12 +1,13 @@
 """A command line tool for extracting text and images from PDF and output it to plain text, html, xml or tags."""
 import argparse
 import sys
-
 import six
 
 import pdfminer.high_level
 import pdfminer.layout
 from pdfminer.image import ImageWriter
+
+logging.basicConfig()
 
 
 def extract_text(files=[], outfile='-',